+++ CHANGES.rst
@@ -34,6 +34,7 @@ CHANGELOG
 **New features**
 
 * Add layer switcher to show satellite background (fixes #123)
+* Add layer switcher to hide POIs in trek detail map (fixes #125)
 * Show detail background when zooming on main map. Disable by setting
   ``TILES_MAIN_MAXZOOM`` to -1.